+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Improvements
 - Adding tooltips to toolbar buttons
 - "Update ?" column renamed to "Upgrade ?"
+- Management panel title renamed
 
 ## [0.2.1] - 2019-06-24
 ### Features