+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.2.2]
+## [0.3.0]
+### Features
+- Allows app uninstall by right clicking it.
+
 ### Improvements
 - Adding tooltips to toolbar buttons
 - "Update ?" column renamed to "Upgrade ?"